update detector materials *important info in commit*
unfortunately it's nontrivial to tell what which orientation in the geometries what material1 and material2 correspond to. even if I were able to tell, ensuring the normals for each triangle are correct (the inner and outer material distinction is defined by the triangle normal) is not something I can guarantee (nor do I think the CAD software we use can). thus I made the simplifying assumption here that simply liquid xenon is the material on up against both sides of each triangle (i.e. `material1=material2='lxe'`). so e.g. instead of seeing the spool wall as a hunk of dense steel, a photon will think the spool is an inner and outer wall of steel separated by LXe. I think this assumption is okay if we do not expect transmission through any of the materials in our geometry. in our surface definitions, steel, gold, and ceramic all have 0% transmission. so then any actual "transmissions" will be through LXe.

if we want to incorporate any filters/glass/anything that's not optically opaque this assumption won't generally hold.