move back to using correct materials *important info in commit msg*
in a previous commit I noted that it was really hard to tell which side of all of the triangles in the solid correspond to being 'inside' or 'outside'. I ended up plotting both sides with different triangles in all cases, material1 referred to the inside of the 3D material (so for a flange, it's the solid metal), and material2 referred to the outside of the material (i.e. the boundary between the surface of the solid with the medium it's submerged in). So to model 3D materials correctly, material1 will be the actual material of the object (same as surface), and material2 will be the medium (e.g. xenon).